+++ R/use_roxyglobals.R
@@ -16,7 +16,10 @@ use_roxyglobals <- function() {
     paste0(utils::packageName(), "::", substitute(global_roclet))
   ))
 
+  # use global_roclet
   options_set_roxygen(options)
+
   # ensure roxyglobals options
+  options_set_filename(options_get_filename())
   options_set_unique(options_get_unique())
 }